Move common code from CreditCard into PaymentSource
Based largely on this commit from mamhoff:

- https://github.com/spree/spree/pull/6831/commits/a3b66ae
